Moving vampire config into slayer category
