+++ src/main/java/at/hannibal2/skyhanni/features/rift/VampireSlayerFeatures.kt
@@ -6,6 +6,7 @@ import at.hannibal2.skyhanni.data.TitleUtils
 import at.hannibal2.skyhanni.events.EntityClickEvent
 import at.hannibal2.skyhanni.events.LorenzTickEvent
 import at.hannibal2.skyhanni.events.withAlpha
+import at.hannibal2.skyhanni.features.rift.everywhere.RiftAPI
 import at.hannibal2.skyhanni.mixins.hooks.RenderLivingEntityHelper
 import at.hannibal2.skyhanni.test.GriffinUtils.drawWaypointFilled
 import at.hannibal2.skyhanni.utils.EntityUtils.getAllNameTagsInRadiusWith
@@ -30,7 +31,7 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 class VampireSlayerFeatures {
 
-    private val config get() = SkyHanniMod.feature.rift.vampireSlayerConfig
+    private val config get() = SkyHanniMod.feature.slayer.vampireSlayerConfig
     private val entityList = mutableListOf<EntityLivingBase>()
     private val taggedEntityList = mutableListOf<Int>()
     private val username = Minecraft.getMinecraft().session.username
